refactor: `AddressSearchResult` 엔티티로 분리
