+++ FoodDiary/Domain/Sources/Repository/AddressSearchRepository.swift
@@ -5,17 +5,6 @@
 
 import Foundation
 
-/// 주소 검색 결과
-public struct AddressSearchResult: Equatable, Sendable {
-    public let placeName: String
-    public let roadAddress: String
-
-    public init(placeName: String, roadAddress: String) {
-        self.placeName = placeName
-        self.roadAddress = roadAddress
-    }
-}
-
 /// 주소 검색 Repository 프로토콜
 public protocol AddressSearchRepository: Sendable {
     /// 키워드로 주소 검색
